Move fixture loading to common conftest.py
